+++ ADMF/functions/Invoke-AdmfDomain.ps1
@@ -86,6 +86,7 @@
 	
 	begin
 	{
+		Reset-DomainControllerCache
 		$parameters = $PSBoundParameters | ConvertTo-PSFHashtable -Include Server, Credential
 		$originalArgument = Invoke-PreCredentialProvider @parameters -ProviderName $CredentialProvider -Parameter $parameters -Cmdlet $PSCmdlet
 		try { $dcServer = Resolve-DomainController @parameters -Confirm:$false }